move scan and adv data to own file, starts type annotations, scanner waitable is now iterable
